fold doc_comment into hover
